Remove cron functionality
There is no useful daemonized version of this image anymore. If you
want it to be run in a cron job, then use the host's cron to perform
the one-time actions. (There isn't a compelling reason to keep this
container running all the time simply to run cron.)

Additional changes:

* Renamed the `no-cron` command `sync` (for consistency with s3cmd).

* Renamed the `delete` command `del` (for consistency with s3cmd).

* The `del` command now passes $PARAMS to s3cmd (i.e. if you want
  to use --dry-run).